refactor: split run_clock orchestrator into runtime_* siblings
run_clock.py had grown to 1356 lines spanning seven unrelated concerns
despite being documented as a "thin orchestrator." Extract the clearly-
separable pieces into sibling modules and re-export them from run_clock
so existing `run_clock.X` imports and test `patch()` sites keep working:

- runtime_log.py        shared _log
- runtime_state.py      RuntimeState container
- runtime_store.py      load/save persisted runtime state
- runtime_telemetry.py  append/prune/daily_telemetry_path
- runtime_theme.py      auto-dark, resolve_effective_theme, midnight reset
- runtime_quiet.py      in_quiet_hours + _display_quiet_image
- runtime_actions.py    action_{skip,unskip,theme,quiet,rerender} + gate

Each action function in runtime_actions.py uses lazy `import run_clock`
internally (same pattern web_server.py already uses) so tests that
patch `run_clock.X` still affect the call path from either caller.

run_clock.py drops from 1356 to ~820 lines; main loop, render pipeline,
button/web wiring, and signal/shutdown lifecycle stay put. No behaviour
change — full 1781-test suite passes. One test patch target updated
(run_clock.dt -> runtime_telemetry.dt) where the patched symbol moved.
